+++ webknossos/tests/dataset/test_attachments.py
@@ -165,7 +165,7 @@ def test_copy_layer(tmp_upath: UPath) -> None:
     )
 
     copy_dataset = Dataset(tmp_upath / "test_copy", voxel_size=(10, 10, 10))
-    copy_layer = copy_dataset.add_copy_layer(seg_layer).as_segmentation_layer()
+    copy_layer = copy_dataset.add_layer_as_copy(seg_layer).as_segmentation_layer()
 
     assert (
         copy_layer.attachments.meshes[0].path
@@ -293,7 +293,7 @@ def test_remote_layer(tmp_upath: UPath) -> None:
     )
 
     copy_dataset = Dataset(tmp_upath / "test_copy", voxel_size=(10, 10, 10))
-    copy_layer = copy_dataset.add_remote_layer(seg_layer).as_segmentation_layer()
+    copy_layer = copy_dataset.add_layer_as_ref(seg_layer).as_segmentation_layer()
 
     assert copy_layer.attachments.meshes[0].path == mesh_path
     assert copy_layer._properties.attachments.meshes is not None
@@ -366,7 +366,7 @@ def test_add_attachments(tmp_upath: UPath) -> None:
         "meshfile_4-4-1",
         data_format=AttachmentDataFormat.Zarr3,
     )
-    seg_layer.attachments.add_attachments(mesh)
+    seg_layer.attachments.add_attachment_as_ref(mesh)
     assert seg_layer._properties.attachments.meshes is not None
     assert seg_layer._properties.attachments.meshes[0].path == "./seg/meshes/meshfile"
     assert seg_layer.attachments.meshes[0].name == "meshfile_4-4-1"
@@ -384,7 +384,7 @@ def test_add_copy_attachments(tmp_upath: UPath) -> None:
         "meshfile_4-4-1",
         data_format=AttachmentDataFormat.Zarr3,
     )
-    seg_layer.attachments.add_copy_attachments(mesh)
+    seg_layer.attachments.add_attachment_as_copy(mesh)
     assert seg_layer._properties.attachments.meshes is not None
     assert seg_layer.attachments.meshes[0].name == "meshfile_4-4-1"
     # path has changed based on the name
@@ -404,7 +404,7 @@ def test_add_copy_attachments(tmp_upath: UPath) -> None:
         "main",
         data_format=AttachmentDataFormat.Zarr3,
     )
-    seg_layer.attachments.add_copy_attachments(segment_index)
+    seg_layer.attachments.add_attachment_as_copy(segment_index)
     assert seg_layer._properties.attachments.segment_index is not None
     assert (
         seg_layer._properties.attachments.segment_index.path